# apiDocumentation.yaml
swagger: '2.0'
info:
  description: ADES CA1
  version: 1.0.0
  title: ADES
  # put the contact info for your development or API team
  contact:
    email: zhengwei.20@ichat.sp.edu.sg

  license:
    name: Website URL
    url: https://lzwades.herokuapp.com
paths:
...
  /authenticate/login/callback:
    post:
      summary: user login with paypal
      operationId: /login/callback
      description: |
        user login with paypal
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: query
        name: code
        required: true
        description: paypal login code
        type: string
      responses:
        200:
          description: return status
          schema:
            $ref: '#/definitions/UserRegisterResult'
        401:
          description: login failed
        500:
          description: Internal Server Error
...
  /authenticate/google:
    get:
      summary: user login with google
      operationId: google
      description: |
        user login with google
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: query
        name: code
        description: google login code
        required: true
        type: string
      responses:
        301:
          description: redirect to index page with the token
        500:
          description: Internal Server Error
...
  /twofa/secretDetail:
    post:
      summary: user enable 2fa
      operationId: secretDetail
      description: |
        user enable 2fa
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: 2fakey
        description: 2fa key
        required: true
        schema: 
          $ref: '#/definitions/TwoFA'
      - in: header
        name: token
        description: user login token
        required: true
        type: string
      responses:
        200:
          description: return message
        401:
          description: token verify error
        500:
          description: Internal Server Error
  /twofa/getSecret:
    post:
      summary: user get secret key
      operationId: getSecret
      description: |
        user get secret key
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: header
        name: token
        description: user login token
        required: true
        type: string
      responses:
        200:
          description: return message
        401:
          description: token verify error
        500:
          description: Internal Server Error
definitions:
...
  UserRegisterResult:
    type: object
    required:
    - token
    properties:
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MywibmFtZSI6IkxpbHkiLCJpYXQiOjE2Mzc5MDI3MjIsImV4cCI6MTYzNzk4OTEyMn0.sybmhllBjRafuPKMlyxnTmuCs2JToI0oF85sC9gmXIg
...
  TwoFA:
    type: object
    required:
    - uid
    - secretkey
    properties:
      uid:
        type: string
        example: 1
      secretkey:
        type: string
        example: qwertyuiop0987654321

# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /lzwades/ades/1.0.0
schemes:
 - https